requirements.txt
annotated-types
anyio
beautifulsoup4
cachetools
certifi
charset-normalizer
colorama
coqpit
decorator
distro
et_xmlfile
exceptiongroup
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-genai
googleapis-common-protos
grpcio
grpcio-status
h11
html5lib
httpcore
httplib2
httpx
idna
ImageHash
imageio
imageio-ffmpeg
langdetect
lxml
moviepy
numpy
oauthlib
opencv-python
openpyxl
packaging
pandas
pillow
proglog
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
pyparsing
pytesseract
python-dateutil
python-dotenv
python-slugify
pytz
PyWavelets
requests
requests-oauthlib
rsa
scipy
six
sniffio
soupsieve
tenacity
text-unidecode
tqdm
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
webencodings
websockets
xlrd
yt-dlp
